check for negative stack

+++ lib/fizzy/types.hpp
@@ -328,7 +328,7 @@ struct Element
 /// https://webassembly.github.io/spec/core/binary/modules.html#code-section
 struct Code
 {
-    unsigned max_stack_depth = 0;
+    size_t max_stack_depth = 0;
 
     uint32_t local_count = 0;
